Pequenas correções de código
