+++ README.md
@@ -1,2 +1,3 @@
-# garoa_status
+# Garoa Status
 
+Este é o código da API para o sensor de porta que está presente no Garoa HC. Ela serve para indicar quando há membros no local ou ainda para verificar se o local encontra-se aberto ou fechado.
