+++ requirements.txt
@@ -1,0 +1,1 @@
+protobuf==3.3.0
